backend/requirements.txt
fastapi
uvicorn
groq
google-generativeai
supabase
python-dotenv
python-multipart
httpx
pydantic
